feat(www): bake in the ruled-table export section
Drops the CTA buttons, the copy width caps, and aligns all five export
destinations on one row grid so available and planned share a column.
